requirements.txt
Django==5.0.2
gunicorn==21.2.0
djangorestframework==3.14.0